a few minor prompt edits
